+++ src/index.ts
@@ -1,0 +1,3 @@
+export { DragScrollModule } from './ngx-drag-scroll.module';
+export { DragScrollComponent } from './ngx-drag-scroll';
+export { DragScrollItemDirective } from './ngx-drag-scroll-item';